ajuste_tratamento_erros_e_exceptions

+++ src/modules/Empreendimentos/repositories/empreendimentos.repository.ts
@@ -17,7 +17,11 @@ export class EmpreendimentoRepository implements EmpreendimentoRepositoryInterfa
     }
 
     async retornarTodosEmpreendimentos(): Promise<Empreendimento[]> {
-        return await this.repo.find();
+        const retorno = await this.repo.find();
+        if (!retorno.length) {
+            throw new NotFoundException('Nenhum empreendimento encontrado');
+        }
+        return retorno;
     }
 
     async retornaEmpreendimentoPorId(id: number) : Promise<Empreendimento> {
@@ -29,11 +33,22 @@ export class EmpreendimentoRepository implements EmpreendimentoRepositoryInterfa
     }
 
     async editarEmpreendimento(id: number, data: Partial<Empreendimento>) {
+        const existente = await this.repo.findOne({ where: { id } });
+        if (!existente) {
+            throw new NotFoundException('Empreendimento não encontrado');
+        }
+
         await this.repo.update(id, data);
         const retorno = await this.retornaEmpreendimentoPorId(id);
         return retorno!;
     }
+
     async excluirEmpreendimento(id: number) {
+        const existente = await this.repo.findOne({ where: { id } });
+        if (!existente) {
+            throw new NotFoundException('Empreendimento não encontrado');
+        }
         await this.repo.delete(id);
+        return { message: `Empreendimento removido com sucesso` };
     }
 }
